Split the core picker menu providers into separate classes.
